+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+### 0.0.3 - Monday 12th October, 2020
+
+-   Added RPC functionality
+-   Reworked the client identification and pubsub features to be extensions on top of RPC
+-   Added HubClient library for client-side support
+-   Enabled HubClient to be used on both Node.js and the browser.
+
 ### 0.0.2 - Thursday 1st October, 2020
 
 -   Fixed a bug where publishing a message to a channel without any subscribers caused an error